requirements.txt
# Apache Airflow - using a version compatible with Python 3.13
apache-airflow>=2.10.0

# Database providers
apache-airflow-providers-postgres>=5.0.0

# Data processing
pandas>=1.5.0

# Testing
pytest>=7.0.0

# Development tools
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
pytest-watch>=4.0.0
